requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.8
djangorestframework>=3.13.1,<4
django-rest-knox>=4.2.0, <5
Markdown>=3.3.7
django-filter>=22.1
Pillow>=9.2,<10